requirements.txt
pandas
numpy
openpyxl
yfinance
requests
urllib3
xgboost
scikit-learn
optuna
streamlit
plotly
kaleido
tqdm
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
